Add files via upload
This is a musical keyboard that can be played using a mouse and/or computer keyboard. Each colored circle represents a note on a standard piano keyboard between C and B. Clicking on any of the colored circles will play a corresponding note and a short ripple animation. C can also be played by pressing the key "z". C# corresponds to "s", and so on, with a gap in the middle row due to the lack of E#. To play an octave lower, press and hold 4. To play an octave higher, press and hold 6. The variable in the upper left corner will reflect which octave you are currently playing, though you will need to continue holding the number key for the audio to be affected.